requirements.txt
Pillow==8.0.1
pygame==2.0.1
PyTMX==3.24